Move flight routers to a separate file
