+++ app.js
@@ -10,6 +10,7 @@ var passport = require('passport');
 
 var indexRouter = require('./routes/indexRouter');
 var loginRouter = require('./routes/loginRouter');
+var flightRouter = require('./routes/flightRouter');
 
 var app = express();
 
@@ -41,6 +42,7 @@ app.use(passport.session());
 
 app.use('/', indexRouter);
 app.use('/auth', loginRouter);
+app.use('/flight', flightRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
